+++ spark-jdbc/src/main/scala/de/dimajix/training/spark/jdbc/WeatherData.scala
@@ -4,12 +4,7 @@ import java.sql.Timestamp
 import java.util.Locale
 
 import org.apache.spark.sql.Row
-import org.apache.spark.sql.types.FloatType
-import org.apache.spark.sql.types.IntegerType
-import org.apache.spark.sql.types.StringType
-import org.apache.spark.sql.types.StructField
-import org.apache.spark.sql.types.StructType
-import org.apache.spark.sql.types.TimestampType
+import org.apache.spark.sql.types._
 import org.joda.time.DateTimeZone
 import org.joda.time.format.DateTimeFormat
 
@@ -28,9 +23,9 @@ object WeatherData extends Serializable {
         StructField("usaf", StringType, false) ::
         StructField("wban", StringType, false) ::
         StructField("air_temperature_quality", IntegerType, false) ::
-        StructField("air_temperature", FloatType, true) ::
+        StructField("air_temperature", DoubleType, true) ::
         StructField("wind_speed_quality", IntegerType, false) ::
-        StructField("wind_speed", FloatType, true) ::
+        StructField("wind_speed", DoubleType, true) ::
         Nil
     )
   }
@@ -41,9 +36,9 @@ object WeatherData extends Serializable {
     val usaf = row.substring(4,10)
     val wban = row.substring(10,15)
     val airTemperatureQuality = row.charAt(92).toInt - '0'.toInt
-    val airTemperature = row.substring(87,92).toFloat/10
+    val airTemperature = row.substring(87,92).toDouble/10
     val windSpeedQuality = row.charAt(69) - '0'.toInt
-    val windSpeed = row.substring(65,69).toFloat/10
+    val windSpeed = row.substring(65,69).toDouble/10
 
     Row(date,time,timestamp,usaf,wban,airTemperatureQuality,airTemperature,windSpeedQuality,windSpeed.toFloat)
   }
